pyproject.toml
[tool.poetry]
name = "aws-gatk-stack"
version = "0.1.0"
description = "CDK Stack for Deploying GATK on AWS"
authors = ["Chris Friedline <chris@diamondage.com>",
"Michael DeRan <mderan@diamondage.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"aws_cdk.core" = "1.95.2"
"aws_cdk.aws_s3" = "*"
"aws_cdk.aws_ec2" = "*"
"aws_cdk.aws_batch" = "*"
"aws_cdk.aws_cloudwatch" = "*"
"aws_cdk.aws_cloudwatch_actions" = "*"
"aws_cdk.aws_ecr" = "*"
"aws_cdk.aws_ecr_assets" = "*"
"aws_cdk.aws_fsx" = "*"
"aws_cdk.aws_iam" = "*"
"aws_cdk.aws_logs" = "*"
"aws_cdk.aws_s3_assets" = "*"
"aws_cdk.aws_secretsmanager" = "*"
boto3 = "^1.17.32"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
isort = "^5.0.6"
black = "^20.8b1"
pre-commit = "^2.6.0"
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
